move types into types
